+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to the "vscode-kusto-workbench" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [4.6.0] - 2026-05-01
+
+* Connections in Connection Manager are now listed alphabetically (d'oh).
+* Adding / removing favorites in Connection Manager gives you a chance to enter a friendly name for the connection (d'oh).
+
 ## [4.5.2] - 2026-04-30
 
 * Added support for VS Code v1.118.1 which contained breaking (for our use cases) changes